refactor(mobile): move shims to platform folder
