+++ apps/mobile/src/platform/crypto-shims.ts
@@ -1,18 +1,10 @@
-import { Platform } from 'react-native'
-import * as Application from 'expo-application'
-const originalFetch = global.fetch
-global.fetch = (url, options = {}) => {
-  options.headers = {
-    ...options.headers,
-    'User-Agent': `SafeMobile/${Platform.OS === 'ios' ? 'iOS' : 'Android'}/${Application.nativeApplicationVersion}/${Application.nativeBuildVersion}`,
-    Origin: 'https://app.safe.global',
-  }
-  return originalFetch(url, options)
-}
-
 import { install } from 'react-native-quick-crypto'
 install()
 
+/**
+ * Override crypto functions in ethers.js to use react-native-quick-crypto
+ * https://docs.ethers.org/v6/cookbook/react-native/
+ */
 import { ethers } from 'ethers'
 
 import crypto from 'react-native-quick-crypto'
@@ -25,14 +17,17 @@ ethers.computeHmac.register((algo, key, data) => {
   return crypto.createHmac(algo, key).update(data).digest()
 })
 
+// @ts-ignore
 ethers.pbkdf2.register((passwd, salt, iter, keylen, algo) => {
   return crypto.pbkdf2Sync(passwd, salt, iter, keylen, algo)
 })
 
 ethers.sha256.register((data) => {
+  // @ts-ignore
   return crypto.createHash('sha256').update(data).digest()
 })
 
 ethers.sha512.register((data) => {
+  // @ts-ignore
   return crypto.createHash('sha512').update(data).digest()
 })
